FEATURE: rename utils folder to types
