+++ src/components/DetailsPageMovieCard.tsx
@@ -1,4 +1,4 @@
-import type { MovieDetails } from "../utils/interface";
+import type { MovieDetails } from "../types/interface";
 import { Image } from "./ui/Image";
 
 export const DetailsPageMovieCard = ({
