+++ src/components/ui/AboutMovie.tsx
@@ -1,4 +1,4 @@
-import type { MovieDetails } from "../../utils/interface";
+import type { MovieDetails } from "../../types/interface";
 
 export const AboutMovie = ({
   description,
